Added new unit tests for ValueCriterion. Fixed a bug: the slope was used as Integer although the DataType of slope is Fractional.

+++ src/main/java/edu/kit/rose/model/YamlProject.java
@@ -125,7 +125,7 @@ class YamlProject {
     private int laneCount;
     @SuppressWarnings("unused")
     @JsonProperty("Steigung")
-    private int slope;
+    private double slope;
     @SuppressWarnings("unused")
     @JsonProperty("Ballungsraum")
     private String conurbation;
